+++ R/req-error.R
@@ -19,7 +19,7 @@
 #' @examples
 #' # Performing this request usually generates an error because httr2
 #' # converts HTTP errors into R errors:
-#' req <- request(test_server()) %>%
+#' req <- request(example_url()) %>%
 #'   req_url_path("/status/404")
 #' try(req %>% req_perform())
 #' # You can still retrieve it with last_response()