+++ R/req-template.R
@@ -19,7 +19,7 @@
 #' @returns A modified HTTP [request].
 #' @export
 #' @examples
-#' httpbin <- request(test_server())
+#' httpbin <- request(example_url())
 #'
 #' # You can supply template parameters in `...`
 #' httpbin %>% req_template("GET /bytes/{n}", n = 100)
@@ -29,7 +29,7 @@
 #' httpbin %>% req_template("GET /bytes/{n}")
 #'
 #' # Existing path is preserved:
-#' httpbin_test <- request(test_server()) %>% req_url_path("/test")
+#' httpbin_test <- request(example_url()) %>% req_url_path("/test")
 #' name <- "id"
 #' value <- "a3fWa"
 #' httpbin_test %>% req_template("GET /set/{name}/{value}")