+++ NEWS.md
@@ -1,5 +1,8 @@
 # httr2 (development version)
 
+* `req_url_path()` and `req_url_path_append()` can handle `NULL` or empty `...`
+  (@mgirlich, #177).
+
 # httr2 0.2.2
 
 * `curl_translate()` can now handle curl copied from Chrome developer tools