+++ NEWS.md
@@ -2,6 +2,9 @@
 
 * Respect custom content type for `req_body_json()` (@mgirlich, #189).
 
+* `req_perform()` now has an `error_call` argument and communicates more clearly
+  where the error occurred (@mgirlich, #187).
+
 # httr2 0.2.2
 
 * `curl_translate()` can now handle curl copied from Chrome developer tools