+++ NEWS.md
@@ -3,6 +3,9 @@
 * `req_url_path()` and `req_url_path_append()` can handle `NULL` or empty `...`
   (@mgirlich, #177).
 
+* `req_perform()` now has an `error_call` argument and communicates more clearly
+  where the error occurred (@mgirlich, #187).
+
 # httr2 0.2.2
 
 * `curl_translate()` can now handle curl copied from Chrome developer tools